split up webpack code and component code
