+++ components/Gallery/specs.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/get-specs";
