+++ components/Gallery/src/Gallery.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta } from "@storybook/addon-docs/blocks";
 
-import { Gallery } from "./Gallery";
+import { Gallery } from ".";
 import notes from "../README.md";
 
 <Meta title="Gallery" parameters={{ notes }} />
