+++ example/Example.mdx
@@ -52,7 +52,8 @@ Buttons allow users to take actions, and make choices, with a single tap.
 ## Types
 
 <Row>
-  <div>
+
+<div>
 
 
 ### Side by side (column size flexible)
@@ -63,12 +64,13 @@ Description of type and breakdown of usage.
 - two
 - three
 
-  </div>
+</div>
 
 
-  <Canvas>
-    <Story id="button--basic" />
-  </Canvas>
+<Canvas>
+  <Story id="button--basic" />
+</Canvas>
+
 </Row>
